+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Add `query_id` to `SQLQueryStatus` events to improve query tracing and debugging
 
 ### Fixes
-- Fix `hard_deletes: invalidate` incorrectly invalidating active records in snapshots (thanks @Zurbste!) ([#1281](https://github.com/databricks/dbt-databricks/issues/1281)) 
+- Fix `hard_deletes: invalidate` incorrectly invalidating active records in snapshots (thanks @Zurbste!) ([#1281](https://github.com/databricks/dbt-databricks/issues/1281))
+- Fix serverless Python model environment configuration: use `environment_version` instead of deprecated `client` field. Users can now specify custom environment versions via `python_job_config.environments`. ([#1277](https://github.com/databricks/dbt-databricks/issues/1277)) 
 
 ## dbt-databricks 1.11.3 (Dec 5, 2025)